Move programming language methods into static Papyros methods
