+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dodona/papyros",
-  "version": "0.1.43",
+  "version": "0.1.51",
   "private": false,
   "homepage": ".",
   "devDependencies": {
