+++ src/App.ts
@@ -3,7 +3,6 @@ import {
     DEFAULT_SERVICE_WORKER, MAIN_APP_ID
 } from "./Constants";
 import { Papyros } from "./Papyros";
-import { plFromString } from "./ProgrammingLanguage";
 import { InputMode } from "./InputManager";
 
 Papyros.configureInput(false, `/${DEFAULT_SERVICE_WORKER}`).then(success => {
@@ -19,7 +18,8 @@ Please use a modern version of Chrome, Safari, Firefox, ...`;
 
 async function startPapyros(): Promise<Papyros> {
     const urlParams = new URLSearchParams(window.location.search);
-    const language = plFromString(urlParams.get("language") || DEFAULT_PROGRAMMING_LANGUAGE);
+    const language = Papyros.toProgrammingLanguage(urlParams.get("language") ||
+        DEFAULT_PROGRAMMING_LANGUAGE)!;
     const locale = urlParams.get("locale") || DEFAULT_LOCALE;
     return Papyros.fromElement({
         standAlone: true,
